# cgroup1: Reduce allocations when reading uints from files