In profiling data from production environments showed that ~20% of all CPU time is spent in the `readUint` function.

https://pprof.me/b97e475/

Upon closer examination we found that a fair amount of time was spent in the `runtime.makeslice` functions. Which suggested that reducing allocations of byte buffers used to read files from disk should yield an reduction in CPU usage.

This patch implements reusing byte buffers for reading files from the cgroupfs into memory, which yielded a ~13% decrease.

```
$ benchstat old.txt new.txt
name           old time/op    new time/op    delta
Readuint64-10    7.93µs ± 2%    6.89µs ± 1%  -13.16%  (p=0.008 n=5+5)

name           old alloc/op   new alloc/op   delta
Readuint64-10      840B ± 0%      120B ± 0%  -85.71%  (p=0.008 n=5+5)

name           old allocs/op  new allocs/op  delta
Readuint64-10      5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.008 n=5+5)
```

And a ~13% decrease of what was originally ~20%, means this should reduce the CPU cycles spent by about 2% overall, but since this also removes some more pressure from garbage collection, the real saving is likely higher, but will only be able to be observed when rolled out in production environments.

We did this as part of a series of live streams where we attempt to optimize popular and widely deployed open-source software in an attempt to reduce the world wide CPU cycles. This was episode number 4: https://www.youtube.com/watch?v=7uz1Zecv1HM