@brancz 2023-04-06
Closing in favor of https://github.com/containerd/cgroups/pull/278